refactor: Move county list to its own package
