+++ app/src/main/java/com/ravigarbuja/covidinfo/di/viewModelModule.kt
@@ -1,6 +1,6 @@
 package com.ravigarbuja.covidinfo.di
 
-import com.ravigarbuja.covidinfo.ui.countries.CountryListViewModel
+import com.ravigarbuja.covidinfo.ui.country.list.CountryListViewModel
 import com.ravigarbuja.covidinfo.ui.main.MainViewModel
 import com.ravigarbuja.covidinfo.ui.summary.SummaryViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
