+++ app/src/main/java/com/ravigarbuja/covidinfo/ui/country/list/CountryListViewModel.kt
@@ -1,8 +1,9 @@
-package com.ravigarbuja.covidinfo.ui.countries
+package com.ravigarbuja.covidinfo.ui.country.list
 
 import androidx.lifecycle.MutableLiveData
 import com.ravigarbuja.covidinfo.base.BaseViewModel
 import com.ravigarbuja.covidinfo.data.model.Country
+import com.ravigarbuja.covidinfo.ui.country.list.CountryListNavigator
 import com.ravigarbuja.covidinfo.util.SingleLiveEvent
 
 class CountryListViewModel : BaseViewModel<CountryListNavigator>(){
